CP6-83 Made modifications based on new information
Modified the channel remapping stub to allow for conjagate flipping with
the new ChannelRemapping struct.
Modified functions that use the ChannelRemapping struct to work with the
new version.
Changed the DFT to an inplace DFT to save copying the endtire signal
data vector.
Removed the downsampling function as it is no longer needed because it
will be done in the channel remapping function now, this will be more
efficant.