CP6-133 Added explicit setup and teardown to unit test framework.
TestModule is now more like a test module factory. It instantiates a TestModule::Impl instance.
TestModule::Impl is polymorphic and enables arbitrary setup and teardown of test modules.
As a (intended) side effect, this allows one to persist arbitrary state for test cases to access. This was previously not practical.
Updated all existing unit tests to work with the new TestModule.