CP6-97 Fixed startup error and frequency channels
Main was deadlocking as the app configuration was not being sent to the secondary nodes in the correct order.
Nodes print a description of what they functions they are doing at any given time.
Fixed the amount of frequency channels in antenna configuration - Was using all of the channels in the metadata, as opposed to files present in the input directory. Updated tests to reflect this change.