CP6-97 Missing voltage files now indicate an error
Added check to see if frequency channels in metadata match the channel voltage files in input directory. If they don't match, this is indicated as an error.
Added public getFrequencyChannels() to metadata file reader.
Changed error return code to 78, since it is easier to distinguish from 1 as an error code. Garrawarla has 78 computing nodes, hence the 78.

Removed a forgotten manual error test in Main.